Feature Request to add BigInteger NumberReturnType support for JsonPathConfig
In order to validate any json integers, you currently need to know the
type ahead of time.  If that's not possible, say in the case of an
integration test on live dynamic data, integer values should be able to
always return the same type.  The common type for integers of any size
is BigInteger.