Provide a method to reset the ThreadLocal in ConfigurableJsonSlurper
This ThreadLocal is static and this is causing issues with environments
with dynamic class loaders as it is keeping a reference to the class
loader, leading to class loader leaks.

Ideally, we should get rid of the class loader but I understand it's
there for compatibility reasons so the next best thing is to provide a
way to reset it so that it can be called without having to rely on nasty
reflection tricks.